+++ Server/Configuration/Config.cs
@@ -91,43 +91,3 @@ public class Config
         }
     }
 }
-
-/// <summary>
-/// Server-specific configuration
-/// </summary>
-public class ServerConfig
-{
-    public int Port { get; set; } = 1701;
-    public bool ShowDebugInConsole { get; set; } = false;
-}
-
-/// <summary>
-/// Database configuration
-/// </summary>
-public class DatabaseConfig
-{
-    public string GameDataFile { get; set; } = "gamedata.db";
-    public string LogDataFile { get; set; } = "gamedata-log.db";
-}
-
-/// <summary>
-/// Logging configuration
-/// </summary>
-public class LoggingConfig
-{
-    public bool EnableConsoleLogging { get; set; } = true;
-    public bool EnableFileLogging { get; set; } = true;
-    public string GameLogFile { get; set; } = "logs/game.log";
-    public string DebugLogFile { get; set; } = "logs/debug.log";
-    public string LogLevel { get; set; } = "Info";
-    public int MaxLogFiles { get; set; } = 5;
-}
-
-/// <summary>
-/// Scripting engine configuration
-/// </summary>
-public class ScriptingConfig
-{
-    public int MaxCallDepth { get; set; } = 100;
-    public int MaxExecutionTimeMs { get; set; } = 5000; // 5 seconds default
-}
