+++ Server/Configuration/ServerConfig.cs
@@ -1,0 +1,10 @@
+namespace CSMOO.Server.Configuration;
+
+/// <summary>
+/// Server-specific configuration
+/// </summary>
+public class ServerConfig
+{
+    public int Port { get; set; } = 1701;
+    public bool ShowDebugInConsole { get; set; } = false;
+}
