+++ Server/Database/Models/FunctionParameter.cs
@@ -1,0 +1,11 @@
+namespace CSMOO.Server.Database.Models;
+
+/// <summary>
+/// Represents a parameter for a user-defined function
+/// </summary>
+public class FunctionParameter
+{
+    public string Name { get; set; } = string.Empty;
+    public string Type { get; set; } = "object";
+    public string DefaultValue { get; set; } = string.Empty;
+}
