+++ Server/Database/Models/GameFunction.cs
@@ -57,13 +57,3 @@ public class GameFunction
     /// </summary>
     public DateTime ModifiedAt { get; set; } = DateTime.UtcNow;
 }
-
-/// <summary>
-/// Represents a parameter for a user-defined function
-/// </summary>
-public class FunctionParameter
-{
-    public string Name { get; set; } = string.Empty;
-    public string Type { get; set; } = "object";
-    public string DefaultValue { get; set; } = string.Empty;
-}
