+++ Server/Scripting/ScriptEngine.cs
@@ -96,312 +96,3 @@ public class ScriptEngine
         }
     }
 }
-
-/// <summary>
-/// Global variables and functions available to scripts
-/// </summary>
-public class ScriptGlobals
-{
-    public Player? Player { get; set; }
-    public CommandProcessor? CommandProcessor { get; set; }
-    public ScriptObjectManager ObjectManager { get; set; } = new ScriptObjectManager();
-    public ScriptWorldManager WorldManager { get; set; } = new ScriptWorldManager();
-    public ScriptPlayerManager PlayerManager { get; set; } = new ScriptPlayerManager();
-    public ScriptHelpers? Helpers { get; set; }
-
-    /// <summary>
-    /// Send a message to the current player
-    /// </summary>
-    public void Say(string message)
-    {
-        if (Helpers != null)
-        {
-            Helpers.Say(message);
-        }
-        else
-        {
-            CommandProcessor?.SendToPlayer(message);
-        }
-    }
-
-    /// <summary>
-    /// Send a message to a specific player
-    /// </summary>
-    public void notify(Player targetPlayer, string message)
-    {
-        if (Helpers != null)
-        {
-            Helpers.notify(targetPlayer, message);
-        }
-        else
-        {
-            CommandProcessor?.SendToPlayer(message, targetPlayer.SessionGuid);
-        }
-    }
-
-    /// <summary>
-    /// Send a message to all players in the current room
-    /// </summary>
-    public void SayToRoom(string message, bool excludeSelf = true)
-    {
-        Helpers?.SayToRoom(message, excludeSelf);
-    }
-
-    /// <summary>
-    /// Get online players
-    /// </summary>
-    public List<Player> GetOnlinePlayers()
-    {
-        return Helpers?.GetOnlinePlayers() ?? new List<Player>();
-    }
-
-    /// <summary>
-    /// Resolve an object name to an object ID
-    /// </summary>
-    public string? ResolveObject(string objectName)
-    {
-        return Helpers?.ResolveObject(objectName);
-    }
-
-    /// <summary>
-    /// Get an object by ID
-    /// </summary>
-    public GameObject? GetObject(string objectId)
-    {
-        return Helpers?.GetObject(objectId);
-    }
-
-    /// <summary>
-    /// Find an object by name
-    /// </summary>
-    public string? FindObjectByName(string name)
-    {
-        return Helpers?.FindObjectByName(name);
-    }
-
-    /// <summary>
-    /// Get a property from an object
-    /// </summary>
-    public object? GetProperty(string objectId, string propertyName)
-    {
-        var obj = Database.ObjectManager.GetObject(objectId);
-        if (obj == null) return null;
-        return Database.ObjectManager.GetProperty(obj, propertyName)?.RawValue;
-    }
-
-    /// <summary>
-    /// Set a property on an object
-    /// </summary>
-    public void SetProperty(string objectId, string propertyName, object value)
-    {
-        Helpers?.SetProperty(objectId, propertyName, value);
-    }
-
-    /// <summary>
-    /// Move an object to a new location
-    /// </summary>
-    public bool MoveObject(string objectId, string destinationId)
-    {
-        return Helpers?.MoveObject(objectId, destinationId) ?? false;
-    }
-
-    /// <summary>
-    /// Show the current room
-    /// </summary>
-    public void ShowRoom()
-    {
-        Helpers?.ShowRoom();
-    }
-
-    /// <summary>
-    /// Look at an object
-    /// </summary>
-    public void LookAtObject(string target)
-    {
-        Helpers?.LookAtObject(target);
-    }
-
-    /// <summary>
-    /// Show player's inventory
-    /// </summary>
-    public void ShowInventory()
-    {
-        Helpers?.ShowInventory();
-    }
-
-    /// <summary>
-    /// Find an item in player's inventory
-    /// </summary>
-    public GameObject? FindItemInInventory(string itemName)
-    {
-        return Helpers?.FindItemInInventory(itemName);
-    }
-
-    /// <summary>
-    /// Find an item in the current room
-    /// </summary>
-    public GameObject? FindItemInRoom(string itemName)
-    {
-        return Helpers?.FindItemInRoom(itemName);
-    }
-
-    /// <summary>
-    /// Create a new object instance
-    /// </summary>
-    public string CreateObject(string className, string? location = null)
-    {
-        var objectClass = DbProvider.Instance.FindOne<ObjectClass>("objectclasses", c => c.Name == className);
-        if (objectClass == null) throw new ArgumentException($"Class '{className}' not found");
-        var newObject = Database.ObjectManager.CreateInstance(objectClass.Id, location);
-        return newObject.Id;
-    }
-
-    /// <summary>
-    /// Find objects by class name
-    /// </summary>
-    public List<string> FindObjects(string className)
-    {
-        var objectClass = DbProvider.Instance.FindOne<ObjectClass>("objectclasses", c => c.Name == className);
-        if (objectClass == null) return new List<string>();
-        return [.. Database.ObjectManager.FindObjectsByClass(objectClass.Id).Select(obj => obj.Id)];
-    }
-
-    /// <summary>
-    /// Call a function on an object (e.g., CallFunction("player", "getName"), CallFunction("system", "display_login"))
-    /// </summary>
-    public object? CallFunction(string objectRef, string functionName, params object?[] parameters)
-    {
-        if (Player == null)
-            throw new InvalidOperationException("No player context available for function calls.");
-
-        var objectId = FunctionResolver.ResolveObjectReference(objectRef, Player.Id, Player.Location?.Id ?? "");
-        if (objectId == null)
-        {
-            throw new ArgumentException($"Object '{objectRef}' not found.");
-        }
-
-        var function = FunctionResolver.FindFunction(objectId, functionName);
-        if (function == null)
-        {
-            throw new ArgumentException($"Function '{functionName}' not found on object '{objectRef}'.");
-        }
-
-        var engine = new UnifiedScriptEngine();
-        return engine.ExecuteFunction(function, parameters, Player, CommandProcessor, objectId);
-    }
-
-    /// <summary>
-    /// Convenience method to call functions on the system object
-    /// </summary>
-    public object? System(string functionName, params object?[] parameters)
-    {
-        return CallFunction("system", functionName, parameters);
-    }
-
-    /// <summary>
-    /// Convenience method to call functions on the calling player
-    /// </summary>
-    public object? Me(string functionName, params object?[] parameters)
-    {
-        return CallFunction("player", functionName, parameters);
-    }
-
-    /// <summary>
-    /// Convenience method to call functions on the current room
-    /// </summary>
-    public object? Here(string functionName, params object?[] parameters)
-    {
-        return CallFunction("here", functionName, parameters);
-    }
-
-    /// <summary>
-    /// Call a function on a specific object by DBREF
-    /// </summary>
-    public object? Object(int dbref, string functionName, params object?[] parameters)
-    {
-        return CallFunction($"#{dbref}", functionName, parameters);
-    }
-
-    /// <summary>
-    /// Call a function on a class
-    /// </summary>
-    public object? Class(string className, string functionName, params object?[] parameters)
-    {
-        return CallFunction($"class:{className}", functionName, parameters);
-    }
-}
-
-/// <summary>
-/// Script-safe wrapper for ObjectManager
-/// </summary>
-public class ScriptObjectManager
-{
-    public object? GetProperty(string objectId, string propertyName)
-    {
-        var obj = ObjectManager.GetObject(objectId);
-        if (obj == null) return null;
-        return Database.ObjectManager.GetProperty(obj, propertyName)?.RawValue;
-    }
-
-    public void SetProperty(string objectId, string propertyName, object value)
-    {
-        var obj = ObjectManager.GetObject(objectId);
-        if (obj != null)
-        {
-            Database.ObjectManager.SetProperty(obj, propertyName, new BsonValue(value));
-        }
-    }
-
-    public void MoveObject(string objectId, string? newLocation)
-    {
-        Database.ObjectManager.MoveObject(objectId, newLocation);
-    }
-
-    public List<string> GetObjectsInLocation(string locationId)
-    {
-        return [.. Database.ObjectManager.GetObjectsInLocation(locationId).Select(obj => obj.Id)];
-    }
-}
-
-/// <summary>
-/// Script-safe wrapper for WorldManager
-/// </summary>
-public class ScriptWorldManager
-{
-    public string? GetStartingRoom()
-    {
-        return Database.WorldManager.GetStartingRoom()?.Id;
-    }
-
-    public List<string> GetAllRooms()
-    {
-        return [.. Database.WorldManager.GetAllRooms().Select(room => room.Id)];
-    }
-
-    public void CreateExit(GameObject fromRoom, GameObject toRoom, string direction, string returnDirection)
-    {
-        Database.WorldManager.CreateExit(fromRoom, toRoom, direction, returnDirection);
-    }
-
-    public string CreateSimpleItem(string name, string shortDesc, string longDesc, string? locationId = null)
-    {
-        return Database.WorldManager.CreateSimpleItem(name, shortDesc, longDesc, locationId).Id;
-    }
-}
-
-/// <summary>
-/// Script-safe wrapper for PlayerManager
-/// </summary>
-public class ScriptPlayerManager
-{
-    public List<string> GetOnlinePlayerNames()
-    {
-        return [.. PlayerManager.GetOnlinePlayers().Select(p => p.Name)];
-    }
-
-    public string? GetPlayerLocation(string playerName)
-    {
-        var player = DbProvider.Instance.FindOne<Player>("players", p => p.Name == playerName);
-        return player?.Location?.Id;
-    }
-}
