+++ Server/Scripting/ScriptObject.cs
@@ -270,203 +270,3 @@ public class ScriptObject : DynamicObject
         return name ?? shortDesc ?? $"Object #{obj.DbRef}";
     }
 }
-
-/// <summary>
-/// Factory for creating ScriptObject instances with natural syntax support
-/// </summary>
-public class ScriptObjectFactory
-{
-    private readonly Player _currentPlayer;
-    private readonly CommandProcessor _commandProcessor;
-    private readonly ScriptHelpers _helpers;
-
-    public ScriptObjectFactory(Player currentPlayer, CommandProcessor commandProcessor, ScriptHelpers helpers)
-    {
-        _currentPlayer = currentPlayer;
-        _commandProcessor = commandProcessor;
-        _helpers = helpers;
-    }
-
-    /// <summary>
-    /// Create a ScriptObject for the given object reference
-    /// Supports: "me", "here", "system", "#123", object names, etc.
-    /// </summary>
-    public dynamic? GetObject(string objectReference)
-    {
-        var objectId = _helpers.ResolveObject(objectReference);
-        if (objectId == null) return null;
-        
-        return new ScriptObject(objectId, _currentPlayer, _commandProcessor, _helpers);
-    }
-
-    /// <summary>
-    /// Create a ScriptObject for a direct object ID
-    /// </summary>
-    public dynamic? GetObjectById(string objectId)
-    {
-        var obj = ObjectManager.GetObject(objectId);
-        if (obj == null) return null;
-        return new ScriptObject(objectId, _currentPlayer, _commandProcessor, _helpers);
-    }
-}
-
-/// <summary>
-/// Enhanced script globals with natural object syntax support
-/// </summary>
-public class EnhancedScriptGlobals : ScriptGlobals
-{
-    private ScriptObjectFactory? _objectFactory;
-
-    /// <summary>
-    /// Initialize the object factory
-    /// </summary>
-    public void InitializeObjectFactory()
-    {
-        // Check for Helpers in base, derived class, or UnifiedScriptGlobals
-        ScriptHelpers? helpers = null;
-        Database.Player? dbPlayer = null;
-        
-        if (this is UnifiedScriptGlobals unifiedGlobals)
-        {
-            helpers = unifiedGlobals.Helpers;
-            // Convert GameObject back to Database.Player if needed
-            dbPlayer = (Database.Player?)unifiedGlobals.Player ?? 
-                      DbProvider.Instance.FindById<Database.Player>("players", ((Database.Player?)unifiedGlobals.Player)?.Id ?? "");
-        }
-        // VerbScriptGlobals merged: handle legacy ThisObjectId
-        else if (this is UnifiedScriptGlobals legacyVerbGlobals && !string.IsNullOrEmpty(legacyVerbGlobals.ThisObjectId))
-        {
-            helpers = legacyVerbGlobals.Helpers;
-            dbPlayer = (Database.Player?)legacyVerbGlobals.Player;
-        }
-        else
-        {
-            helpers = Helpers;
-            dbPlayer = Player as Database.Player;
-        }
-        
-        if (dbPlayer != null && CommandProcessor != null && helpers != null)
-        {
-            _objectFactory = new ScriptObjectFactory(dbPlayer, CommandProcessor, helpers);
-        }
-    }
-
-    /// <summary>
-    /// Get object with natural syntax: GetObject("player") or GetObject("#123")
-    /// Usage: var player = GetObject("me"); player.Name = "NewName"; player:sayHello();
-    /// </summary>
-    public new dynamic? GetObject(string objectReference)
-    {
-        return _objectFactory?.GetObject(objectReference);
-    }
-
-    /// <summary>
-    /// The player executing the verb
-    /// Usage: player.Name = "NewName"; player.Location = "room_id";
-    /// </summary>
-    public dynamic? player
-    {
-        get
-        {
-            string? playerId = null;
-            if (this is UnifiedScriptGlobals unifiedGlobals)
-            {
-                playerId = ((Database.Player?)unifiedGlobals.Player)?.Id;
-            }
-            else
-            {
-                playerId = Player?.Id;
-            }
-            return _objectFactory?.GetObjectById(playerId ?? "");
-        }
-    }
-
-    /// <summary>
-    /// The object this verb is running on (same as 'this')
-    /// Usage: me.Name = "New Name"; me:someVerb();
-    /// </summary>
-    public dynamic? me
-    {
-        get
-        {
-            if (this is UnifiedScriptGlobals unifiedGlobals)
-            {
-                return _objectFactory?.GetObjectById((unifiedGlobals.This as GameObject)?.Id ?? "");
-            }
-            else if (this is UnifiedScriptGlobals legacyVerbGlobals && !string.IsNullOrEmpty(legacyVerbGlobals.ThisObjectId))
-            {
-                return _objectFactory?.GetObjectById(legacyVerbGlobals.ThisObjectId);
-            }
-            return null;
-        }
-    }
-
-    /// <summary>
-    /// The current room (where the player is)
-    /// Usage: here.Name = "New Room Name"; here:look();
-    /// </summary>
-    public dynamic? here
-    {
-        get
-        {
-            string? location = null;
-            if (this is UnifiedScriptGlobals unifiedGlobals)
-            {
-                location = ((Database.Player?)unifiedGlobals.Player)?.Location?.Id;
-            }
-            else
-            {
-                location = Player?.Location?.Id;
-            }
-            return _objectFactory?.GetObjectById(location ?? "");
-        }
-    }
-
-    /// <summary>
-    /// The system object as a ScriptObject
-    /// Usage: system:who(); system.Name;
-    /// </summary>
-    public dynamic? system => _objectFactory?.GetObject("system");
-
-    /// <summary>
-    /// The object this verb is running on (same as 'me')
-    /// Usage: this.Name = "New Name"; this:someVerb();
-    /// </summary>
-    public dynamic? @this
-    {
-        get
-        {
-            if (this is UnifiedScriptGlobals unifiedGlobals)
-            {
-                return _objectFactory?.GetObjectById((unifiedGlobals.This as GameObject)?.Id ?? "");
-            }
-            else if (this is UnifiedScriptGlobals legacyVerbGlobals && !string.IsNullOrEmpty(legacyVerbGlobals.ThisObjectId))
-            {
-                return _objectFactory?.GetObjectById(legacyVerbGlobals.ThisObjectId);
-            }
-            return null;
-        }
-    }
-
-    /// <summary>
-    /// Get object by DBREF number: obj(123) returns object #123
-    /// Usage: var target = obj(123); target.Name = "NewName";
-    /// </summary>
-    public dynamic? obj(int dbref)
-    {
-        return _objectFactory?.GetObject($"#{dbref}");
-    }
-
-    /// <summary>
-    /// Call a verb on an object with natural syntax
-    /// Usage: CallVerb("player", "getName", "arg1", "arg2")
-    /// Note: It's better to use the object syntax: GetObject("player"):getName("arg1", "arg2")
-    /// </summary>
-    public string CallVerb(string objectReference, string verbName, params object[] args)
-    {
-        dynamic? obj = GetObject(objectReference);
-        if (obj == null) return $"Object '{objectReference}' not found";
-        
-        return obj.CallVerb(verbName, args);
-    }
-}
