--- Server/ServerDatabase.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Text.Json;
-using System.Text.Json.Serialization;
-
-namespace CSMOO.Server;
-
-internal class ServerDatabase
-{
-
-}
-
