+++ Server/Session/SessionInfo.cs
@@ -1,0 +1,6 @@
+using System;
+using CSMOO.Server.Session;
+
+namespace CSMOO.Server.Session;
+
+public record SessionInfo(Guid ClientGuid, IClientConnection Connection);
